Refactor TokensList
