+++ src/dashboard-react-web/src/components/views/Service/ViewZone/Tokens/Tokens.scss
